requirements.txt

yfinance
pandas
matplotlib
fpdf2
requests
textblob
python-dotenv
streamlit
Pillow
